techdebt(parser): split NodeArena mutation APIs by AST family (#9432)
`crates/tsz-parser/src/parser/node_arena.rs` was a 2372-line monolith
that exceeded the §19 hard 2000-line cap and held every `add_*` AST
mutator in one impl block, making parser-ownership boundaries hard to
see.

Convert the file into a directory whose `mod.rs` keeps lifecycle and
parent-pointer orchestration (under 450 lines) and route AST family
mutators through eight new submodules, each contributing its own
`impl NodeArena { ... }`:

- `expressions.rs` — binary/call/access/conditional/literal/template/
  type-assertion/qualified-name/computed-property/spread/expr-with-args
- `declarations.rs` — function/class/interface/type-alias/enum/module/
  variable
- `statements.rs` — block/if/loop/for-in-of/return/switch/try/labeled/
  jump/with
- `types.rs` — type ref/composite/function-type/query/literal/array/
  tuple/wrapped/conditional/infer/operator/indexed-access/mapped/
  literal-type/template-literal/named-tuple/predicate
- `members.rs` — signature/index-signature/property/method/constructor/
  accessor/parameter/type-parameter/decorator/heritage
- `patterns.rs` — binding-pattern/binding-element/property-assignment/
  shorthand-property
- `imports_exports.rs` — import/export decls, clauses, specifiers,
  named imports, import attributes
- `jsx.rs` — JSX element/fragment/attribute/expression/text/namespaced
  nodes

Storage stays centralized on `NodeArena` (definition unchanged in
`parser::node`). The shared helpers `set_parent`, `set_parent_list`,
`set_parent_opt_list`, and `len_u32` move from private `fn` to
`pub(super)` so submodules can call them through `self`. Public method
signatures, bodies, parent-link ordering, push order, and the `clear`
field set are preserved byte-for-byte; the existing `#[cfg(test)] mod
tests` block stays in `mod.rs`.

Structural rule: when the parser builds an AST node through any
`NodeArena::add_*` mutator, the resulting `NodeIndex`, parent links,
extended-info slot, and storage layout are unchanged from before the
split — this is implementation-only partitioning.

Verification: `cargo build -p tsz-parser`, `cargo clippy -p tsz-parser
--all-targets -- -D warnings`, `cargo test -p tsz-parser --lib`
(1031 passed / 0 failed), `python3 scripts/arch/arch_guard.py`, and
`scripts/arch/check-checker-boundaries.sh` all pass on this head.
